requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-dotenv==0.19.0
Pillow==8.3.1
boto3==1.18.0
pytz==2021.1 